+++ bin/ons_csv_to_ctb_json_main.py
@@ -29,6 +29,12 @@ def main():
                         required=True,
                         help='Output directory to write JSON files')
 
+    parser.add_argument('-g', '--geography-file',
+                        type=str,
+                        required=False,
+                        help='Name of CSV file containing category codes and names for geographic '
+                             'variables')
+
     parser.add_argument('-l', '--log_level',
                         type=str,
                         default='INFO',
@@ -44,7 +50,7 @@ def main():
                         level=args.log_level)
 
     # loader is used to load the metadata from CSV files and convert it to JSON.
-    loader = Loader(args.input_dir)
+    loader = Loader(args.input_dir, args.geography_file)
 
     # Build Cantabular variable and dataset objects and write them to a JSON file.
     # A Cantabular variable is equivalent to an ONS classification.
@@ -88,6 +94,7 @@ def build_ctb_variables(classifications, cat_labels):
         ctb_class = {
             'name': mnemonic,
             'label': classification.private['Classification_Label'],
+            'description': classification.private['Variable_Description'],
             'meta': classification,
             'catLabels': cat_labels.get(mnemonic, None)
         }
@@ -117,12 +124,12 @@ def build_ctb_datasets(databases, ctb_variables):
             'Base dataset with metadata for all variables in Welsh',
         ),
         'lang': Bilingual('en', 'cy'),
+        'description': Bilingual(
+            'This is a base dataset containing metadata for all variables used across all '
+            'other datasets. Other datasets include it to avoid duplicating metadata.',
+            'This is the Welsh version of the base dataset containing metadata for all '
+            'variables.'),
         'meta': {
-            'Database_Description': Bilingual(
-                'This is a base dataset containing metadata for all variables used across all '
-                'other datasets. Other datasets include it to avoid duplicating metadata.',
-                'This is the Welsh version of the base dataset containing metadata for all '
-                'variables.'),
             'Source': {
                 'Source_Mnemonic': 'Census2021',
                 'Source_Description': 'The 2021 England and Wales Census',
@@ -138,6 +145,7 @@ def build_ctb_datasets(databases, ctb_variables):
             'name': database_mnemonic,
             'incl': [{'name': 'base', 'lang': Bilingual('en', 'cy')}],
             'label': database.private['Database_Title'],
+            'description': database.private['Database_Description'],
             'lang': Bilingual('en', 'cy'),
             'meta': database,
             'vars': None,