+++ test/test_integration.py
@@ -36,7 +36,8 @@ class TestIntegration(unittest.TestCase):
         file_dir = pathlib.Path(__file__).parent.resolve()
         input_dir = os.path.join(file_dir, 'testdata')
         output_dir = os.path.join(file_dir, 'out')
-        with unittest.mock.patch('sys.argv', ['test', '-i', input_dir, '-o', output_dir]):
+        geo_dir = os.path.join(input_dir, 'geography/geography.csv')
+        with unittest.mock.patch('sys.argv', ['test', '-i', input_dir, '-o', output_dir, '-g', geo_dir]):
             ons_csv_to_ctb_json_main.main()
             with open(os.path.join(output_dir, 'service-metadata.json')) as f:
                 service_metadata = json.load(f)
@@ -47,6 +48,31 @@ class TestIntegration(unittest.TestCase):
             with open(os.path.join(output_dir, 'dataset-metadata.json')) as f:
                 dataset_metadata = json.load(f)
             with open(os.path.join(file_dir, 'expected/dataset-metadata.json')) as f:
+                expected_dataset_metadata = json.load(f)
+            self.assertEqual(dataset_metadata, expected_dataset_metadata)
+
+            with open(os.path.join(output_dir, 'table-metadata.json')) as f:
+                table_metadata = json.load(f)
+            with open(os.path.join(file_dir, 'expected/table-metadata.json')) as f:
+                expected_table_metadata = json.load(f)
+            self.assertEqual(table_metadata, expected_table_metadata)
+
+    def test_no_geography_file(self):
+        """Generate JSON from source CSV and compare it with expected values."""
+        file_dir = pathlib.Path(__file__).parent.resolve()
+        input_dir = os.path.join(file_dir, 'testdata')
+        output_dir = os.path.join(file_dir, 'out')
+        with unittest.mock.patch('sys.argv', ['test', '-i', input_dir, '-o', output_dir]):
+            ons_csv_to_ctb_json_main.main()
+            with open(os.path.join(output_dir, 'service-metadata.json')) as f:
+                service_metadata = json.load(f)
+            with open(os.path.join(file_dir, 'expected/service-metadata.json')) as f:
+                expected_service_metadata = json.load(f)
+            self.assertEqual(service_metadata, expected_service_metadata)
+
+            with open(os.path.join(output_dir, 'dataset-metadata.json')) as f:
+                dataset_metadata = json.load(f)
+            with open(os.path.join(file_dir, 'expected/dataset-metadata-no-geo.json')) as f:
                 expected_dataset_metadata = json.load(f)
             self.assertEqual(dataset_metadata, expected_dataset_metadata)