+++ test/test_related_dataset.py
@@ -20,7 +20,7 @@ class TestRelatedDataset(unittest.TestCase):
         with unittest.mock.patch('builtins.open', conditional_mock_open('Related_Datasets.csv',
                 read_data = build_test_file(HEADERS, rows))):
             with self.assertRaisesRegex(ValueError, expected_error):
-                Loader(INPUT_DIR).load_dataset_to_related(['DS1', 'DS2', 'DS3'])
+                Loader(INPUT_DIR, None).load_dataset_to_related(['DS1', 'DS2', 'DS3'])
 
     def test_required_fields(self):
         for field in REQUIRED_FIELDS: