+++ test/test_topic.py
@@ -22,7 +22,7 @@ class TestTopic(unittest.TestCase):
         with unittest.mock.patch('builtins.open', conditional_mock_open('Topic.csv',
                 read_data = build_test_file(HEADERS, rows))):
             with self.assertRaisesRegex(ValueError, expected_error):
-                Loader(INPUT_DIR).topics
+                Loader(INPUT_DIR, None).topics
 
     def test_required_fields(self):
         for field in REQUIRED_FIELDS: