+++ test/test_variable_keyword.py
@@ -20,7 +20,7 @@ class TestVariableKeyword(unittest.TestCase):
         with unittest.mock.patch('builtins.open', conditional_mock_open('Variable_Keyword.csv',
                 read_data = build_test_file(HEADERS, rows))):
             with self.assertRaisesRegex(ValueError, expected_error):
-                Loader(INPUT_DIR).load_variable_to_keywords(['VAR1', 'VAR2'])
+                Loader(INPUT_DIR, None).load_variable_to_keywords(['VAR1', 'VAR2'])
 
     def test_required_fields(self):
         for field in REQUIRED_FIELDS: